Tudor Networks of Power (Mini Project)
This mini-project is a capstone for the Python-based activities in our course. The idea is to give you an opportunity to delve into a real network dataset that you have not seen before, starting with less structure than in the labs and homeworks.

We will focus on an epistolary network - that is, a network of correspondence. This network is derived from an analysis of the Tudor State Papers, which is a historical archive of official papers and personal letters from the Tudor Era (a historical period in England and Wales spanning roughly 1485-1603).  This network has been the subject of a recent study called Tudor Networks of Power, and the epistolary network we will investigate was published along with that book. Researchers used the Tudor State Papers to create a network of individuals and letters sent between these individuals; these letters can be modeled as a network.

You can find the dataset here: 

https://github.com/tudor-networks-of-power/code/tree/main/TNP_DATA

You will primarily use letter_edgelist.tsv. (If you decide to do the Extra Credit, then you will also use people_labels.tsv to look up which ID corresponds to each historic figure.)